Allow some multi-word units, remove US-/UK- prefix
Accepts units like "survey foot" and "british mile" now, instead of the
previous solution, which was prefixes to adjust between international
and old units such as US- and UK-. These prefixes were confusing because
they could be used on anything, not just customary units of
length. USgallon makes no sense, for example, because it is multiplying
the volume of a gallon by the difference between a survey foot and an
international foot.