Automatically derive molar mass from formulas.
This commit does not recognize any radioactive elements in formulas
that rarely if ever appear as a part of compound, i.e. every one
except Bi, Th, U, Pu. This is done to be able to recognize some
common abbreviations in organic chemistry: Me, Et, Ac, Ph, some
of which clash with transuranics.