add prolactin’s biological activity “density” (µg/mIU)
This commit defines international units (IU) of biological activity,
as well as the biological activity “density” of the current reference
preparation of prolactin, to make it easier to convert prolactin
levels between ng/mL (as preferred in the US) and mIU/L (as preferred
in Australia).

I can’t find a formal term for the relationship between biological
activity and mass in reference preparations, so I’ve chosen “density”
by analogy with linear_density, area_density, etc. For example:

> 256 mIU/L * biological_activity_density of prolactin -> ng/mL
4096/335, approx. 12.22686 nanogram / milliliter (density)