Add more number output formats (#180)
Adds `to frac` and `to sci` to explicitly force the use of these
representations. Also adds `to eng` which behaves similar to the default
behavior, except rounds to every third power of 10 instead of every
power of 10. I might make the output format configurable and set the
engineering one as the default in a future version.

While working on this I found a bug in the current version of rink.
`surveyfoot to digits` is incorrect by 3 orders of magnitude. I fixed
that in this PR.