Markup for doc strings (#177)
Makes Rink parse out the links and date strings when using the token
format machinery. This is a new syntax in the definitions file with
`<abc>` for links and `#abc#` for dates.

Replaced ansi_term with nu-ansi-term because the former was
unmaintained. The latter comes with the ability to set hyperlinks.