feat(core): Add "relative dates"
I was looking at your project and really liked it, especially for some quick
date calculations, such as `#2023-05-03# - #2024-06-22# in months;weeks;days`
which turned out to be so simple to do compared to spinning up either a
spreadsheet, node.js, python etc. What I was missing though was, as you might've
guessed, a feature to specify relative dates such as `today:-1`... This way the
above calculation could've been just:
 `#2023-05-03# - #today# in months;weeks;days`! Hence this PR.

As for the code, it is a tad copy-pasty, but the MVP is there. I constrained it
to "just" weeks and days, as to keep it a bit more sane. I hope you like the
idea!