Try replacing codecov (#215)
Codecov has been flaky for a while and now it's started failing
entirely. For some reason, it says there is no coverage data, despite a
lcov.info file being fully successfully uploaded to their site. I can
download it back from them and it's a non-empty lcov file.

So I'm going to try to scrape together something that doesn't rely on
proprietary SaaS and see how it goes.

This PR adds something based on the Python `diff_cover` tool with some
extra information shoved in, then posted as a comment on the PR.

Eventually I'd like to gain back the ability to see the total % coverage
change vs the merge base, but this seems to be almost impossible with
Github Actions.